1) Send notification to trainer when group session is approved/rejected and trainee join trainer's 1-1 training. 2) Update doBond personalsession db update 3) Added check to determine is for user or group session when admin want to reject group session for doReject.php.

+++ administrator/doReject.php
@@ -1,18 +1,40 @@
 <?php
 require('../database/dbconfig.php');
 
-$email = $_POST['reject_userid'];
+// check if form submitted is for user approval or group session approval
+preg_match("/@+/", $_POST['reject_userid'], $output_array);
 
-$sql = "DELETE FROM userapproval WHERE email = '$email'";
-$query = $conn->prepare($sql);
-$stmt = $query->execute();
+if (count($output_array) == 1) {
+    $email = $_POST['reject_userid'];
 
-//reject for groupsession
-$id = $_POST['reject_userid'];
-$sql3 = "UPDATE groupsession SET status = 'Rejected' WHERE id = '$id'";
-$query4 = $conn->prepare($sql3);
-$stmt3 = $query4->execute();
+    $sql = "DELETE FROM userapproval WHERE email = '$email'";
+    $query = $conn->prepare($sql);
+    $stmt = $query->execute();
 
-header('Location: user-approval.php');
+    header('Location: user-approval.php');
+    
+} else { //reject for groupsession
+    $id = $_POST['reject_userid'];
+    $sql5 = "SELECT * FROM groupsession WHERE id = '$id'";
+    $query5 = $conn->prepare($sql5);
+    $stmt5 = $query5->execute();
+    $row22 = $req22->fetch(PDO::FETCH_ASSOC);
+    $dateformat = $row22['date'];
+    $starttime = $row22['startTime'];
+    $roomtypeid = $row22['roomTypeID'];
+    $trainerEmail = $row22['trainerEmail'];
+
+    // send a notification to the trainer
+    $msg = "Your group session on $dateformat at $starttime has been rejected.";
+    $sql2 = "INSERT into notificationlog (message, userEmail, readStatus) VALUES ('$msg', '$trainerEmail', '0')";
+    $query2 = $conn->prepare($sql2);
+    $stmt2 = $query2->execute();
+    
+    $sql3 = "UPDATE groupsession SET status = 'Rejected' WHERE id = '$id'";
+    $query4 = $conn->prepare($sql3);
+    $stmt3 = $query4->execute();
+    
+    header('Location: user-approval.php');
+}
 exit();
